# dist/openapi.yaml
openapi: 3.0.0
info:
  title: Recipes api
  description: API for recipes project
  version: 1.0.1
servers:
# local:
- url: "http://127.0.0.1:3000"
# remote:
# - url: "http://132.72.65.211"
  description: SwaggerHub API Auto Mocking hey
...
paths:
  /register:
    post:
      tags:
      - Auth
      summary: Add new user to the server.
      description: Add new user to the server.
      requestBody:
        description: Inventory item to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        "201":
          description: A new user has been added
        "400":
          description: wrong input parameters
        "409":
          description: Username taken.
  /login:
    post:
      tags:
      - Auth
      summary: Login to the system.
      description: An existing user logins to the system
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginScheme'
        required: true
      responses:
        "200":
          description: |
            User was found in the DB and authenitcated successfully. The session ID is returned in a cookie named `session`.  You should include this cookie in subsequent requests
          headers:
            Set-Cookie:
              style: simple
              explode: false
              schema:
                type: string
                example: session=87s97sdhgf279293f8w0e9
        "401":
          description: Incorrect username or password
...
  /recipes/{recipeId}:
    get:
      tags:
      - Recipes
      summary: Return preview details on recipe by his ID
      description: Return preview details on recipe by his ID
      parameters:
      - name: recipeId
        in: path
        description: recipe's id
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: data about the recipe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipePreview'
        "404":
          description: recipeId is not exists
...
  /search:
    get:
      tags:
        - Recipes
      summary: Search for recipes with according to query.
      description: Search for recipe
      parameters:
        - name: searchQuery
          in: query
          description: The query to search
          required: true
          style: form
          explode: true
          schema:
            type: string
            example: cake
        - name: resultsSize
          in: query
          description: how many results to return
          required: true
          style: form
          explode: true
          schema:
            type: integer
            default: 5
            maximum: 15
            minimum: 5
            multipleOf: 5
            example: 10
        - name: IsGlutenFree
          in: query
          description: cuisine tags, SEPERATED BY COMMAS
          required: false
          style: form
          explode: true
          schema:
            type: boolean
        - name: dietTags
          in: query
          description: diet tags, SEPERATED BY COMMAS
          required: false
          style: form
          explode: true
          schema:
            type: string
            enum:
            - vegan
            - vegetarian
            example: Vegetarian or Vegan only
        
      responses:
        '200':
          description: Return the recipes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MultiplyRecipes'
        '400':
          description: Error.
  /random:
    get:
      tags:
        - Recipes
      summary: Return 3 random recipie ID's.
      description: Get three random recipes.
      responses:
        '200':
          description: Return the recipes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LastViewed'
        '400':
          description: Error.
components:
  schemas:
...
    User:
      type: object
      properties:
        username:
          maxLength: 8
          minLength: 3
          type: string
          description: username conatins between 3 to 8 letters only
          example: my_name
        firstname:
          type: string
          description: user's first name
          example: my_first_name
        lastname:
          type: string
          description: user's last name
          example: my_last_name
        country:
          type: string
          description: user's country
          example: Israel
        password:
          maxLength: 10
          minLength: 5
          type: string
          description: password conatins between 5 to 10 chars, at least one of them
            is a number and a special char
          format: password
          example: my_name@12
        email:
          pattern: email
          type: string
          description: user's email
          example: my_name12@gmail.com
        # image-url:
        #   pattern: URL
        #   type: string
        #   description: URL for the user's image
        #   example: https://cloudinary.com/
      description: Model containing user info
    LoginScheme:
      type: object
      properties:
        username:
          maxLength: 8
          minLength: 3
          type: string
          description: username conatins letters only
        password:
          maxLength: 10
          minLength: 5
          description: password conatins at least one number and one special char
          format: password
      example:
        username: my_name
        password: my_name@12
    LastViewed:
      maxItems: 3
      minItems: 3
      uniqueItems: true
      type: array
      description: Array of 3 recipes
      items:
        $ref: '#/components/schemas/RecipePreview'
    RecipePreview:
      type: object
      properties:
        id:
          type: integer
          description: recipe identifier
          example: 102
        title:
          type: string
          description: recipe name
          example: Simple Macaroni and Cheese
        image:
          type: string
          description: recipe image
          format: URL
          example: https://spoonacular.com/recipeimages/655455-566x370.jpg
        readyInMinutes:
          type: integer
          description: the time takes to make this recipe
          example: 45
        popularity:
          type: integer
          description: how many likes got this recipe
          example: 5987
        vegetarian:
          type: boolean
          description: indication if the recipe is vegetarian
          example: false
        vegan:
          type: boolean
          description: indication if the recipe is vegan
          example: false
        glutenFree:
          type: boolean
          description: indication if the recipe is gluten free
          example: false
...
    MultiplyRecipes:
      type: array
      description: unlimited amount of reduced recipes
      items:
        $ref: '#/components/schemas/RecipePreview'